#626 Implement usage of caching during database calls
- Added applicationsInAppCollection cache
