+++ src/main/java/ai/elimu/rest/v1/project/AppCollectionRestController.java
@@ -1,24 +1,13 @@
 package ai.elimu.rest.v1.project;
 
 import ai.elimu.dao.AppCollectionDao;
-import ai.elimu.dao.ApplicationDao;
-import ai.elimu.dao.ApplicationVersionDao;
 import ai.elimu.dao.LicenseDao;
-import ai.elimu.model.admin.Application;
-import ai.elimu.model.admin.ApplicationVersion;
-import ai.elimu.model.enums.Locale;
-import ai.elimu.model.gson.admin.ApplicationGson;
-import ai.elimu.model.gson.admin.ApplicationVersionGson;
 import ai.elimu.model.gson.project.AppCollectionGson;
-import ai.elimu.model.project.AppCategory;
 import ai.elimu.model.project.AppCollection;
-import ai.elimu.model.project.AppGroup;
 import ai.elimu.model.project.License;
 import ai.elimu.rest.v1.JavaToGsonConverter;
+import ai.elimu.service.JsonService;
 import com.google.gson.Gson;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import org.apache.log4j.Logger;
 import org.json.JSONArray;
@@ -49,10 +38,7 @@ public class AppCollectionRestController {
     private AppCollectionDao appCollectionDao;
     
     @Autowired
-    private ApplicationDao applicationDao;
-    
-    @Autowired
-    private ApplicationVersionDao applicationVersionDao;
+    private JsonService jsonService;
     
     @RequestMapping(method = RequestMethod.GET)
     public String get(
@@ -109,8 +95,6 @@ public class AppCollectionRestController {
             jsonObject.put("result", "error");
             jsonObject.put("description", "Invalid license");
         } else {
-            Date dateStart = new Date();
-            logger.info("appCollectionDao.read(" + appCollectionId + ") - " + dateStart);
             AppCollection appCollection = appCollectionDao.read(appCollectionId);
             if (appCollection == null) {
                 jsonObject.put("result", "error");
@@ -118,97 +102,14 @@ public class AppCollectionRestController {
             } else {
                 // TODO: verify that the AppCollection matches the one associated with the provided License
                 
-                JSONArray applications = new JSONArray();
-                
-                addInfrastructureApps(applications);
-                
-                for (AppCategory appCategory : appCollection.getAppCategories()) {
-                    for (AppGroup appGroup : appCategory.getAppGroups()) {
-                        for (Application application : appGroup.getApplications()) {
-                            ApplicationGson applicationGson = JavaToGsonConverter.getApplicationGson(application);
-
-                            // Fetch the Application's ApplicationVersions
-                            List<ApplicationVersionGson> applicationVersions = new ArrayList<>();
-                            logger.info("applicationVersionDao.readAll(" + application.getPackageName() + ") - " + new Date());
-                            for (ApplicationVersion applicationVersion : applicationVersionDao.readAll(application)) {
-                                logger.info("applicationVersion: " + applicationVersion.getVersionCode() + " - " + new Date());
-                                ApplicationVersionGson applicationVersionGson = JavaToGsonConverter.getApplicationVersionGson(applicationVersion);
-                                applicationVersions.add(applicationVersionGson);
-                            }
-                            applicationGson.setApplicationVersions(applicationVersions);
-
-                            String json = new Gson().toJson(applicationGson);
-                            applications.put(new JSONObject(json));
-                        }
-                    }
-                }
+                JSONArray applications = jsonService.getApplications(appCollection);
 
                 jsonObject.put("result", "success");
                 jsonObject.put("applications", applications);
             }
-            Date dateEnd = new Date();
-            long duration = dateEnd.getTime() - dateStart.getTime();
-            logger.info("duration: " + (duration / 1000) + " seconds");
         }
         
         logger.info("jsonObject: " + jsonObject);
         return jsonObject.toString();
-    }
-    
-    /**
-     * As AppCollections in Custom Projects do not include all the Applications from the open source 
-     * version, some apps are required to form the basic infrastructure.
-     */
-    private void addInfrastructureApps(JSONArray applications) {
-        // Appstore
-        Application applicationAppstore = applicationDao.readByPackageName(Locale.EN, "ai.elimu.appstore");
-        if (applicationAppstore != null) {
-            ApplicationGson applicationGson = JavaToGsonConverter.getApplicationGson(applicationAppstore);
-
-            // Fetch the Application's ApplicationVersions
-            List<ApplicationVersionGson> applicationVersions = new ArrayList<>();
-            for (ApplicationVersion applicationVersion : applicationVersionDao.readAll(applicationAppstore)) {
-                ApplicationVersionGson applicationVersionGson = JavaToGsonConverter.getApplicationVersionGson(applicationVersion);
-                applicationVersions.add(applicationVersionGson);
-            }
-            applicationGson.setApplicationVersions(applicationVersions);
-
-            String json = new Gson().toJson(applicationGson);
-            applications.put(new JSONObject(json));
-        }
-        
-        // Analytics
-        Application applicationAnalytics = applicationDao.readByPackageName(Locale.EN, "ai.elimu.analytics");
-        if (applicationAnalytics != null) {
-            ApplicationGson applicationGson = JavaToGsonConverter.getApplicationGson(applicationAnalytics);
-
-            // Fetch the Application's ApplicationVersions
-            List<ApplicationVersionGson> applicationVersions = new ArrayList<>();
-            for (ApplicationVersion applicationVersion : applicationVersionDao.readAll(applicationAnalytics)) {
-                ApplicationVersionGson applicationVersionGson = JavaToGsonConverter.getApplicationVersionGson(applicationVersion);
-                applicationVersions.add(applicationVersionGson);
-            }
-            applicationGson.setApplicationVersions(applicationVersions);
-
-            String json = new Gson().toJson(applicationGson);
-            applications.put(new JSONObject(json));
-        }
-        
-        // Custom Launcher
-        Application applicationCustomLauncher = applicationDao.readByPackageName(Locale.EN, "ai.elimu.launcher_custom");
-        if (applicationCustomLauncher != null) {
-            ApplicationGson applicationGson = JavaToGsonConverter.getApplicationGson(applicationCustomLauncher);
-
-            // Fetch the Application's ApplicationVersions
-            List<ApplicationVersionGson> applicationVersions = new ArrayList<>();
-            for (ApplicationVersion applicationVersion : applicationVersionDao.readAll(applicationCustomLauncher)) {
-                ApplicationVersionGson applicationVersionGson = JavaToGsonConverter.getApplicationVersionGson(applicationVersion);
-                applicationVersions.add(applicationVersionGson);
-            }
-            applicationGson.setApplicationVersions(applicationVersions);
-
-            String json = new Gson().toJson(applicationGson);
-            applications.put(new JSONObject(json));
-        }
     }
 }
